Moved functionality from StartMenu to StartBar
